+++ assets/components/modal-footer.component.jsx
@@ -3,7 +3,7 @@ import customPropTypes from '../utilities/component/custom-prop-types';
 import getPassThroughProperties from '../utilities/component/get-pass-through-properties';
 import pureRenderShouldComponentUpdate from '../utilities/pure-render-should-component-update';
 
-class Card extends React.Component {
+class ModalFooter extends React.Component {
   constructor(props) {
     super(props);
   }
@@ -13,14 +13,14 @@ class Card extends React.Component {
   }
 
   getCssClasses() {
-    let cssClasses = ['card'];
+    let cssClasses = ['modal__footer'];
 
     if (this.props.className) {
       cssClasses = cssClasses.concat(this.props.className.split(' '));
     }
 
-    if (this.props.styleType) {
-      cssClasses.push('m-' + this.props.styleType)
+    if (this.props.isActions) {
+      cssClasses.push('is-actions');
     }
 
     return cssClasses;
@@ -30,7 +30,7 @@ class Card extends React.Component {
     return (
       <div
         className={this.getCssClasses().join(' ')}
-        {...getPassThroughProperties(this.props, 'className', 'styleType')}
+        {...getPassThroughProperties(this.props, 'className', 'isActions')}
       >
         {this.props.children}
       </div>
@@ -38,16 +38,16 @@ class Card extends React.Component {
   }
 }
 
-Card.displayName = 'Card';
+ModalFooter.displayName = 'ModalFooter';
 
-Card.propTypes = {
+ModalFooter.propTypes = {
   className: React.PropTypes.string,
-  styleType: customPropTypes.cardStyleTypes
+  isActions: React.PropTypes.bool
 };
 
-Card.defaultProps = {
+ModalFooter.defaultProps = {
   className: null,
-  styleType: null
+  isActions: false
 };
 
-export default Card;
+export default ModalFooter;